fix: clarify naming
